requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.37.0
datasets>=2.14.0
peft>=0.7.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
opencv-python>=4.8.0
pycocotools>=2.0.7
qwen-vl-utils>=0.0.2
